Fix data types in CalculatorTest.java

+++ src/test/java/com/example/calculator/CalculatorTest.java
@@ -10,25 +10,25 @@ public class CalculatorTest {
 
     @Test
     public void testRoot() {
-        double result = calculator.root(9);
-        assertEquals(3, result, 0.001);
+        double result = calculator.root(9.0);
+        assertEquals(3.0, result, 0.001);
     }
 
     @Test
     public void testFactorial() {
-        double result = calculator.factorial(5);
+        int result = calculator.factorial(5.0);
         assertEquals(120, result, 0.001);
     }
 
     @Test
     public void testLog() {
-        double result = calculator.log(5);
+        double result = calculator.log(5.0);
         assertEquals(1.61, result, 0.001);
     }
 
     @Test
     public void testPower() {
-        double result = calculator.power(5, 2);
+        int result = calculator.power(5.0, 2.0);
         assertEquals(25, result, 0.001);
     }
 }
